+++ packages/column-views/src/prepare-units/helpers.ts
@@ -16,6 +16,99 @@ import type { ExtUnit, SectionInfo, StratigraphicPackage } from "./types";
 
 const dt = 0.001;
 
+export type UnitWithDefinedOverlap<T extends BaseUnit> = T & {
+  overlap: UnitOverlapInfo<T>;
+};
+
+type UnitOverlapInfo<T extends BaseUnit> = {
+  column: number;
+  nColumns: number;
+  unitGroup: UnitGroup<T>;
+};
+
+class UnitGroup<T extends BaseUnit> implements StratigraphicPackage {
+  units: T[];
+
+  constructor(units: T[]) {
+    this.units = units;
+  }
+
+  get t_age(): number {
+    return Math.min(...this.units.map((d) => d.t_age));
+  }
+
+  get b_age(): number {
+    return Math.max(...this.units.map((d) => d.b_age));
+  }
+
+  getUnits(): UnitWithDefinedOverlap<T>[] {
+    return assignColumnsToUnitsWithinOverlappingGroup(this);
+  }
+}
+
+function groupUnitsByMutualOverlap<T extends BaseUnit>(
+  units: T[],
+  tolerance: number = dt,
+): UnitGroup<T>[] {
+  /** Group units within a section that have mutual overlap. */
+  const sections = groupUnitsIntoSectionsByOverlap(
+    units,
+    ColumnAxisType.AGE,
+    tolerance,
+  );
+  return sections.map((d) => {
+    return new UnitGroup(d.units);
+  });
+}
+
+const sortByAge = createUnitSorter(ColumnAxisType.AGE);
+
+function assignColumnsToUnitsWithinOverlappingGroup<T extends BaseUnit>(
+  group: UnitGroup<T>,
+): UnitWithDefinedOverlap<T>[] {
+  /** A relatively naïve function to assign columns to units within a mutually overlapping group.
+   * This gets the number of columns that would be needed to accommodate the units in the group.
+   * It then assigns columns to units in the group based on the number of columns needed, with
+   * lower units further to the left.
+   */
+
+  const sortedUnits = [...group.units].sort((a, b) => b.b_age - a.b_age);
+
+  const columns: T[][] = [];
+
+  for (const unit of sortedUnits) {
+    // Find a column that is not already occupied by an overlapping unit
+    let didAddToExistingColumn = false;
+    for (const column of columns) {
+      const lastUnit = column[column.length - 1];
+      if (!unitsOverlap(unit, lastUnit, ColumnAxisType.AGE, dt)) {
+        column.push(unit);
+        didAddToExistingColumn = true;
+        break;
+      }
+    }
+    if (!didAddToExistingColumn) {
+      columns.push([unit]);
+    }
+  }
+
+  const units: UnitWithDefinedOverlap<T>[] = columns.flatMap(
+    (column, columnIndex) =>
+      column.map((unit) => {
+        return {
+          ...unit,
+          overlap: {
+            column: columnIndex,
+            nColumns: columns.length,
+            unitGroup: group,
+          },
+        };
+      }),
+  );
+
+  return units.toSorted(sortByAge);
+}
+
 export function preprocessUnits<T extends UnitLong = UnitLong>(
   section: SectionInfo<T>,
   axisType: ColumnAxisType = ColumnAxisType.AGE,
@@ -23,88 +116,43 @@ export function preprocessUnits<T extends UnitLong = UnitLong>(
   /** Preprocess units to add overlapping units and columns. */
   let units = section.units;
 
-  let divisions = units.map((...args) => extendDivision(...args, axisType));
-  for (let d of divisions) {
-    const overlappingUnits = divisions.filter((u) =>
-      d.overlappingUnits.includes(u.unit_id),
-    );
-
-    // Overlapping columns
-    const columns = overlappingUnits.map((d) => d.column);
-
-    if (columns.includes(d.column)) {
-      let col = 0;
-      // Go through columns to find a better index
-      while (columns.includes(col)) {
-        col++;
-      }
-      d.column ??= col;
-    }
-
-    // If unit overlaps the edges of a section, set the clip positions
-    const [b_pos, t_pos] = getUnitHeightRange(d, axisType);
-    if (axisType == ColumnAxisType.AGE) {
-      if (b_pos > section.b_age) {
-        d.b_clip_pos = section.b_age;
-      }
-      if (t_pos < section.t_age) {
-        d.t_clip_pos = section.t_age;
-      }
-    }
-    if (axisType == ColumnAxisType.DEPTH) {
-      if (b_pos > section.b_pos) {
-        d.b_clip_pos = section.b_pos;
-      }
-      if (t_pos < section.t_pos) {
-        d.t_clip_pos = section.t_pos;
-      }
-    }
-    // if (axisType == ColumnAxisType.HEIGHT) {
-    //   if (b_pos < section.b_pos) {
-    //     d.b_clip_pos = section.b_pos;
-    //   }
-    //   if (t_pos > section.t_pos) {
-    //     d.t_clip_pos = section.t_pos;
-    //   }
-    // }
-  }
-
-  return divisions;
-}
-
-function extendDivision(
-  unit: UnitLong,
-  i: number,
-  divisions: UnitLong[],
-  axisType: ColumnAxisType = ColumnAxisType.AGE,
-): ExtUnit {
-  // TODO: make this configurable
   let tolerance = 0.001; // 1 kyr tolerance for age columns
   if (axisType != ColumnAxisType.AGE) {
     tolerance = 0.01; // 1cm tolerance for height/depth columns
   }
 
-  const overlappingUnits = divisions.filter(
-    (d) =>
-      d.unit_id != unit.unit_id && unitsOverlap(unit, d, axisType, tolerance),
-  );
-  const u_pos = getUnitHeightRange(unit, axisType);
-  const bottomOverlap = overlappingUnits.some((d) => {
-    const d_pos = getUnitHeightRange(d, axisType);
-    // Check if the unit is below the current unit
-    return d_pos[0] < u_pos[0] + dt;
-  });
-  let column = 0;
-  if (overlappingUnits.length == 1) {
-    column = 1;
+  const unitGroups = groupUnitsByMutualOverlap(units, tolerance);
+
+  const divisions: any[] = [];
+  // Serialize all groups' units to a single array
+  for (const group of unitGroups) {
+    divisions.push(...group.getUnits());
   }
 
-  return {
-    ...unit,
-    bottomOverlap,
-    column: unit.column ?? column,
-    overlappingUnits: overlappingUnits.map((d) => d.unit_id),
-  };
+  return divisions.map((d) =>
+    enhanceUnitWithClippingPosition(section, d, axisType),
+  );
+}
+
+export function enhanceUnitWithClippingPosition(section, d, axisType) {
+  const [b_pos, t_pos] = getUnitHeightRange(d, axisType);
+  if (axisType == ColumnAxisType.AGE) {
+    if (b_pos > section.b_age) {
+      d.b_clip_pos = section.b_age;
+    }
+    if (t_pos < section.t_age) {
+      d.t_clip_pos = section.t_age;
+    }
+  }
+  if (axisType == ColumnAxisType.DEPTH) {
+    if (b_pos > section.b_pos) {
+      d.b_clip_pos = section.b_pos;
+    }
+    if (t_pos < section.t_pos) {
+      d.t_clip_pos = section.t_pos;
+    }
+  }
+  return d;
 }
 
 export function groupUnitsIntoSectionsBySectionID<T extends UnitLong>(
@@ -141,9 +189,10 @@ interface WorkingSection {
   heightRange?: [number, number];
 }
 
-export function groupUnitsIntoSectionsByOverlap<T extends UnitLong>(
+export function groupUnitsIntoSectionsByOverlap<T extends BaseUnit>(
   units: T[],
   axisType: ColumnAxisType = ColumnAxisType.AGE,
+  tolerance: number = dt,
 ): SectionInfo<T>[] {
   /** Group units into sections by overlap.
    * This creates "synthetic" sections that correspond to packages bound by scale gaps.
@@ -157,7 +206,7 @@ export function groupUnitsIntoSectionsByOverlap<T extends UnitLong>(
     const heightRange = getUnitHeightRange(unit, axisType);
     let section: WorkingSection | undefined = sectionList.find(
       (s) =>
-        compareAgeRanges(heightRange, s.heightRange) !==
+        compareAgeRanges(heightRange, s.heightRange, tolerance) !==
         AgeRangeRelationship.Disjoint,
     );
     if (section == null) {