+++ packages/column-views/src/unit-details/index.ts
@@ -2,3 +2,5 @@ export * from "./panel";
 export * from "./popover";
 export * from "./modal-panel";
 export * from "./age-range";
+export * from "./strat-names";
+export * from "./identifiers";