+++ packages/form-components/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.0.9] - 2026-05-20 [_changes_](https://github.com/UW-Macrostrat/web-components/compare/@macrostrat/form-components-v1.0.8...@macrostrat/form-components-v1.0.9)
+
+### Patch Changes
+
+- Update blueprintjs dependencies
+  [239c0292](https://github.com/UW-Macrostrat/web-components/commit/239c0292ce143f642cf05ec934f246798d89a54f)
+- Updated dependencies
+  [239c0292](https://github.com/UW-Macrostrat/web-components/commit/239c0292ce143f642cf05ec934f246798d89a54f)
+  - @macrostrat/ui-components@5.0.10
+
 ## [1.0.8] - 2026-05-19 [_changes_](https://github.com/UW-Macrostrat/web-components/compare/@macrostrat/form-components-v1.0.7...@macrostrat/form-components-v1.0.8)
 
 ### Patch Changes